cope with errors from recvfrom (from ig25@fg70.rz.uni-karlsruhe.de (Thomas Koenig))
SCCS-vsn: lib/librpc/rpc/svc_udp.c 1.3

+++ usr/src/lib/librpc/rpc/svc_udp.c
@@ -182,7 +182,7 @@ svcudp_recv(xprt, msg)
 	    0, (struct sockaddr *)&(xprt->xp_raddr), &(xprt->xp_addrlen));
 	if (rlen == -1 && errno == EINTR)
 		goto again;
-	if (rlen < 4*sizeof(u_long))
+	if (rlen < (int)(4*sizeof(u_long)))
 		return (FALSE);
 	xdrs->x_op = XDR_DECODE;
 	XDR_SETPOS(xdrs, 0);
